requirements.txt
azure-storage-blob
pandas
psycopg2-binary